renaming registrar into registry
