+++ app/scripts/lib/ens-ipfs/contracts/registry.js
